requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.158
botocore==1.29.158
bs4==0.0.1
click==8.1.3
gitdb==4.0.10
GitPython==3.1.31
Jinja2==3.1.2
jmespath==1.0.1
jsondiff==2.0.0
jsonschema==4.17.3
mariadb==1.1.7
MarkupSafe==2.1.3
packaging==23.1
pony==0.7.17
PyMySQL==1.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
urllib3==1.26.16
